Admin color schemes: Manually handle RTL and minified versions of the CSS files on registration.
This bypasses WP_Styles entirely, which is much simpler for the moment, given that color schemes bypass WP_Styles for plenty already. The script loader is told to stop thinking of colors.css as an RTL-ified file. The colors-fresh handle, used directly on the login screen, needed to be (even before this commit).

fixes #26316.


git-svn-id: https://develop.svn.wordpress.org/trunk@26780 602fd350-edb4-49c9-b593-d223f7449a82