Menus: Prevent display conflicts in the screen options for Menus when a post type and taxonomy share the same slug.
Prior to this change, both taxonomies and post types added meta boxes to the Menus UI (which in-turn registered screen options) using the same ID formula: `add-{$id}`. This caused a conflict between the two elements, resulting in the rendering of a screen option for only the taxonomy or post type but not both. Moving to a new meta box ID for post types, `add-post-type-{$id}`, removes that conflict.

Note: This change in meta box IDs for post types effectively resets current user settings for the visibility of post types in the Menus UI.

Fixes #33925.


git-svn-id: https://develop.svn.wordpress.org/trunk@35450 602fd350-edb4-49c9-b593-d223f7449a82