Editor: Allow the default font size options to be removed.
Similar to `editor-color-palette`, a theme or plugin should be able to set the theme support for `editor-font-sizes` to an empty array, indicating that no additional font sizes should be available in the block’s ‘Text Settings’. The current conditional was using `empty()`, which was incorrectly causing the default font sizes to be used instead.

Props jorgefilipecosta, chrisvanpatten, SergeyBiryukov.
Fixes #46290.

git-svn-id: https://develop.svn.wordpress.org/trunk@44782 602fd350-edb4-49c9-b593-d223f7449a82