Twenty Twenty-One: Correct `@since` inline documentation tags.
Previously, `1.0.0` was used for the version. This is inconsistent with the versioning used in the other default themes, which use use X.X and not X.X.X.

This change corrects all `1.0.0` occurrences to `1.0`, and adds the additional context of `Twenty Twenty-One` to avoid confusion with `WordPress 1.0`.

Props SergeyBiryukov, poena.
Fixes #51958.

git-svn-id: https://develop.svn.wordpress.org/trunk@49826 602fd350-edb4-49c9-b593-d223f7449a82