Customizer: Bring admin color schemes back into the customizer.
In #50547, the admin color schemes were added to the customizer UI, but at some point specificity changed and the colors were overridden by the defaults. This brings the color schemes back into the customizer, using the button color as a highlight, with the link color for text to keep things readable.

Follow up to [48371].
Props audrasjb, peterwilsoncc.
Fixes #52230.



git-svn-id: https://develop.svn.wordpress.org/trunk@50386 602fd350-edb4-49c9-b593-d223f7449a82