Twenty Twenty-One: Add the px unit to the admin bar height custom property.
The admin bar height custom property is used in `calc` functions, which require consistent unit use, even when the value is zero. A unitless value returns an invalid value for the "height" property, where this variable is used. This also changes the stylelint rule in the theme to allow zero values with a unit, just in custom properties.

Props nico23.
Fixes #52564.



git-svn-id: https://develop.svn.wordpress.org/trunk@50388 602fd350-edb4-49c9-b593-d223f7449a82