Bundled Themes: Update block patterns to match the latest versions of `core/*` blocks.
When using the post editor with a bundled theme active, there will be a number of `console.info` notices printed to the browser’s console.

These notices are caused by block patterns containing outdated and deprecated versions of `core/*` blocks. Before rendering the blocks, they need to be updated, and this process outputs information to the console.

Props ntsekouras.
Fixes #53617.

git-svn-id: https://develop.svn.wordpress.org/trunk@51372 602fd350-edb4-49c9-b593-d223f7449a82