Formatting: Use `safecss_filter_attr` in `wp_get_layout_style()`.
Previously, `safecss_filter_attr` did not allow for `calc` and `var` in parsed CSS.  Because of this, `wp_strip_all_tags` was used in `wp_get_layout_style` to allow for the use of these CSS functions to account for responsive widths.  [50923] added support for both of these CSS functions in KSES, so `safecss_filter_attr` can now be used in `wp_get_layout_style` as originally intended.

Follow-up to [52069].

Props grapplerulrich.
Fixes #55356.

git-svn-id: https://develop.svn.wordpress.org/trunk@52924 602fd350-edb4-49c9-b593-d223f7449a82