Posts, Post Types: Update new `wp_post_class_taxonomies` filter name for consistency.
In [54066], a new filter was added to alter the taxonomies for which class names are generated for a given post type.  At the time, the filter name was prefixed with `wp_`.

For consistency with filters of a similar type, this change updates the filter name to `post_class_taxonomies`.

Follow-up to [54066].

Props SergeyBiryukov.
See #37114.

git-svn-id: https://develop.svn.wordpress.org/trunk@54208 602fd350-edb4-49c9-b593-d223f7449a82