Upgrade/Install: Send language pack updates to `copy_dir()`.
If the `clear_working` flag in `WP_Upgrader::install_package()` is false, the source should not be removed, so `copy_dir()` should be used instead.

Partial updates, like language packs, may want to retain the destination. If the destination exists or has contents, this may be a partial update, and the destination should not be removed, so `copy_dir()` should be used instead.

Follow-up to [55204], [55219], [55220], [55223], [55226].

Props afragen, costdev, swissspidy.
See #57557.

git-svn-id: https://develop.svn.wordpress.org/trunk@55229 602fd350-edb4-49c9-b593-d223f7449a82