I18N: Correct translator comments placement in `wp-admin/edit-comments.php`.
This ensures that the translator comments are displayed for the strings they refer to, e.g. to explain placeholders in comment action messages, and are not displayed for unrelated strings, e.g. "Undo", to avoid confusion.

Includes minor code layout fixes and wrapping a few long lines for better readability.

Follow-up to [35549], [45926], [45932].

Props afercia.
See #57839.

git-svn-id: https://develop.svn.wordpress.org/trunk@55630 602fd350-edb4-49c9-b593-d223f7449a82