Build/Test Tools: Downgrade `grunt-contrib-qunit` dependency.
`grunt-contrib-qunit` was upgraded from version `7.0.1` to `8.0.1` in [56647]. However, this update causes a strange failure when running the build script for the Gutenberg plugin when checked out within a `wordpress-develop` checkout.

This reverts the related change in [56647] and downgrades the dependency back to `7.0.1` until the exact reason for the failure is narrowed down.

Props afercia, kevin940726, antonvlasenko, desrosj.
See #59634, #58863.


git-svn-id: https://develop.svn.wordpress.org/trunk@56944 602fd350-edb4-49c9-b593-d223f7449a82