Tests: Update expectations in `wp_remote_head()` and `wp_remote_get()` tests.
It appears that something has changed on the WP.com side to compress the requested images on the fly, which interfered with the previous expectations in these tests.

This commit uses a direct file URL and updates the expected image size to match the currently returned response.

Follow-up to [139/tests], [31258], [47142].

Props dextorlobo, swissspidy, davidbaumwald, SergeyBiryukov.
See #60865.

git-svn-id: https://develop.svn.wordpress.org/trunk@57903 602fd350-edb4-49c9-b593-d223f7449a82