Customize: Avoid loading the `widgets` component if a block theme is activated.
This changeset adds a check to ensure the current theme is not a block theme before adding the `widgets` component to the Customizer manager. This avoids a PHP warning when previewing/customizing notification emails with a block theme in BuddyPress and potentially on other plugins.

See https://github.com/buddypress/buddypress/pull/215.

Props imath, mukesh27, faisal03, sumitbagthariya16, webcommsat.
Fixes #60236.




git-svn-id: https://develop.svn.wordpress.org/trunk@58376 602fd350-edb4-49c9-b593-d223f7449a82