Role/Capability: Introduce the `current_user_can_for_site()` and `user_can_for_site()` functions.
The `current_user_can_for_site()` function is a replacement for `current_user_can_for_blog()` which is now deprecated. `user_can_for_site()` is a renaming of the `user_can_for_blog()` function which was introduced in [59123]. The intention of this change is to prevent the introduction of a new function which uses the old "blog" naming structure.

Props swissspidy, spacedmonkey, flixos90, johnjamesjacoby

Fixes #45197


git-svn-id: https://develop.svn.wordpress.org/trunk@59198 602fd350-edb4-49c9-b593-d223f7449a82