Themes: Improve performance of applying background image styles in `theme.json`.
The cost of using `WP_Theme_JSON::get_block_nodes()` for this in its original shape was high enough to lead to a performance regression. Therefore this changeset introduces a new option on the method that allows to bypass all logic except for retrieving the node paths, which is much faster and everything that this functionality needs.

Follow up to [58936].

Props mukesh27, flixos90, ramonopoly, joemcgill, andrewserong, swissspidy.
Fixes #61858.


git-svn-id: https://develop.svn.wordpress.org/trunk@59213 602fd350-edb4-49c9-b593-d223f7449a82