Docs: Correct description for `the_time()`, `get_the_time()`, and `get_post_time()`.
This aims to avoid confusion with `the_date()` and `get_the_date()`.

Includes synchronizing the description for `the_weekday()` and `the_weekday_date()`, which have very similar functionality, except that the latter will only output the weekday if the current post's weekday is different from the previous one output.

Follow-up to [59691].

See #51289.

git-svn-id: https://develop.svn.wordpress.org/trunk@59692 602fd350-edb4-49c9-b593-d223f7449a82