Administration: Introduce a lighter background for WP Admin.
This changeset replaces the light grey background color with the white color defined in the Editor Storybook. This change also impacts admin color schemes that previously utilized the default admin background color.
This is an initial implementation of the WordPress design system, aligning with the broader goal of achieving a more consistent and unified design across the administration.

Props karmatosed, audrasjb.
Fixes #62831.




git-svn-id: https://develop.svn.wordpress.org/trunk@59705 602fd350-edb4-49c9-b593-d223f7449a82