Plugins: Improve search box margin behavior in the Add Plugins screen.
This changeset fixes a margin issue in the search input box on the Add New Plugins screen, which was previously breaking below 1138px. Specifically, the top margin was set to 0px, and the overall appearance of the search box was inconsistent between 1000px and 1138px. Now, the margin is consistent across all breakpoints.

Props jomonthomaslobo1, narenin, iflairwebtechnologies, peterwilsoncc, audrasjb, shailu25.
Fixes #61785.




git-svn-id: https://develop.svn.wordpress.org/trunk@59706 602fd350-edb4-49c9-b593-d223f7449a82