Build/Test: Update assertions in  AtomParser_Parse_Test::test_parse_sets_handlers
On some instances of Windows, the assertions seem to find additional nodes. As this test is just about verifying that the handlers get called, not about testing the functionality of the handlers, we can adjust the assertion to look for a minimum number of nodes rather than exact number.

Follow-up to [59062].

Props yogeshbhutkar, hellofromTonya, SergeyBiryukov, coquardcyr, jrf, benniledl, desrosj, jorbin.
Fixes #62110. See #62061.



git-svn-id: https://develop.svn.wordpress.org/trunk@59739 602fd350-edb4-49c9-b593-d223f7449a82