Upgrade/Install: Prevent an unnecessary plugin update check when the plugin update data is up to date.
This ensures the `checked` property is always populated with the latest plugin data before determining whether to perform an update check. Previously this was only populated when the data was already identified as being stale, which could result in a subsequent unnecessary update check when viewing the Plugins screen.

Props siliconforks, bookdude13, pbiron, francina, Cybr, snehapatil02, johnbillion

Fixes #44118, #61055

git-svn-id: https://develop.svn.wordpress.org/trunk@59926 602fd350-edb4-49c9-b593-d223f7449a82