Build/Test: Remove unused ink-docstrap.
jsdoc is not currently in use and this theme for it is very out of date. There are currently a small handful of minor vulnerabilities within transitive dependencies being privately reported by Dependabot that cannot be resolved due to version constraints within ink-docstrap. While this in no way affects WordPress (since this code is not used by WordPress), it does create noise which can be eliminated.

This was first added in [41351] as a part of #41682.

Props desrosj.
Fixes #62935.


git-svn-id: https://develop.svn.wordpress.org/trunk@59935 602fd350-edb4-49c9-b593-d223f7449a82