Editor: Support site preview mode for the Site Editor.
This allows the Site Editor to load a preview of the site as an `IFRAME_REQUEST` by checking for a `wp_site_preview` query param. By setting the `IFRAME_REQUEST` constant when loading the preview, the site's admin bar and some third party plugin overlays will not be shown in the preview.

Props wildworks, joemcgill, peterwilsoncc, presstoke, mamaduka, johnbillion, faisal03.
Fixes #63076.


git-svn-id: https://develop.svn.wordpress.org/trunk@60003 602fd350-edb4-49c9-b593-d223f7449a82