Build/Test Tools: Prevent Dependabot workflow runs in other mirrors.
Currently, workflows are configured to only run for `wordpress-develop` or when pull requests are opened where forks and mirrors are the base repository.

Because a Dependabot configuration is present in the repository, it cannot be turned off for mirrors. This results in workflows running for all Dependabot PRs, which is problematic for private mirrors and needlessly consumes minutes and resources.

Props swissspidy, johnbillion.
See #62221.

git-svn-id: https://develop.svn.wordpress.org/trunk@60080 602fd350-edb4-49c9-b593-d223f7449a82