Build/Test Tools: Adjust how the 6.8 branch is included in matrices.
Because 6.8 is not yet officially released, `6.8` cannot be used in a test matrix.

This removes 6.8 from the Upgrade Testing workflow and changes the value used for the Upgrade Develop Version Testing workflow to `6.8-RC1` to confirm that updating from `6.8-RC1` to `trunk` (which is now `6.9-alpha`) is successful.

Follow-up to [60093].

Props mukesh27.
See #63164.

git-svn-id: https://develop.svn.wordpress.org/trunk@60097 602fd350-edb4-49c9-b593-d223f7449a82