Bundled Themes: Bump `Tested up to` for Twenty Twenty-Three & Twenty Twenty-Four.
An update for these two default themes will not be released in coordination with WordPress 6.8 due because there were no changes to the source code during the release cycle. However, they should be considered tested with 6.8, and this ensures that the next release properly indicates that.

Props shailu25, sabernhardt, desrosj, poena.
See #62990.

git-svn-id: https://develop.svn.wordpress.org/trunk@60158 602fd350-edb4-49c9-b593-d223f7449a82