About Page: Reduce copy on 6.8 about page.
In finalizing the 6.8 announcement post (https://wordpress.org/news/2025/04/cecil/) it was identified that some of the copy on the 6.8 about page is not an accurate description of what is in 6.8. This copy is receiving a game misconduct penalty and automatic ejection. While the coach could bring on a substitute, a line change is the better option here.

Props abcd95, mukesh27, JeffPaul.
Fixes #63323.


git-svn-id: https://develop.svn.wordpress.org/trunk@60183 602fd350-edb4-49c9-b593-d223f7449a82