Build/Test Tools: Test version in additional config files.
Introduces tests to ensure the version numbers in `package-lock.json` (two instances) and `composer.json` match the WordPress version specified in `version.php`.

In pull requests, the `package-lock.json` file is updated automatically to match the version in `package.json`. The new test is intended to ensure the version numbers are correct in production branches.

Fixes #63390.



git-svn-id: https://develop.svn.wordpress.org/trunk@60219 602fd350-edb4-49c9-b593-d223f7449a82