Build/Test Tools: Fix file name references in [60219].
While it's true that package.json uses the format x.y.z, for the tests of other files using the same format it's best practice for the comments to reference the file being tested rather than provide a cool story about an unrelated file.

Cool story.

Follow up to [60219].

See #63390.


git-svn-id: https://develop.svn.wordpress.org/trunk@60221 602fd350-edb4-49c9-b593-d223f7449a82