External Libraries: Upgrade Simple Pie to 1.8.1 (patched).
Upgrades the Simple Pie library to a patched version of [https://github.com/simplepie/simplepie/releases/tag/1.8.1 Simple Pie 1.8.1]. Much of 1.8.1 was included in the 1.8.0 upgrade committed in r59141. The following fixes from the latest release those that remain for this upgrade:

* Fix locator with website missing `Content-Type` header [https://github.com/simplepie/simplepie/pull/891 simplepie/simplepie#891]
* Fix `encode` argument of `SimplePie::strip_htmltags()` [https://github.com/simplepie/simplepie/pull/894 simplepie/simplepie#894]

A caching fix not included in Simple Pie 1.8.1 is also included in this upgrade, see [https://github.com/simplepie/simplepie/pull/883 simplepie/simplepie#883].

A caching test for `fetch_feed()` is introduced in this pull request to ensure that the caching patch is included in future upgrades of the library.

Props kaygee79, oglekler, SergeyBiryukov, peterwilsoncc.
Fixes #63717.


git-svn-id: https://develop.svn.wordpress.org/trunk@60490 602fd350-edb4-49c9-b593-d223f7449a82