Customize: Enable Customizer live previews of Custom CSS in Block Themes.
In a Block Theme, the Custom CSS from the Customizer is inserted into the global styles, before the global style's Custom CSS. In order to update the Customizer's Custom CSS inside of the global styles, milestone CSS comments are added in the preview so that the live preview logic can do the replacement.

Follow-up to [55192], [58703].
See #57536, #61395.

Props westonruter, wildworks, peterwilsoncc, rollybueno, SirLouen, poojapadamad, rafiq91, audrasjb.
Fixes #63589.


git-svn-id: https://develop.svn.wordpress.org/trunk@60522 602fd350-edb4-49c9-b593-d223f7449a82