Charset: Add explanatory note about what consitutes “valid” UTF-8.
This patch adds a clarifying note about what constitutes a valid UTF-8 byte stream. This was brought up in review as a potentially ambiguous term, so a link to the spec has been provided to fix the behavior to the standard.

Developed in https://github.com/WordPress/wordpress-develop/pull/9716
Discussed in https://core.trac.wordpress.org/ticket/38044

Follow-up to [60630].

Props dmsnell, agulbra.
See #38044.


git-svn-id: https://develop.svn.wordpress.org/trunk@60702 602fd350-edb4-49c9-b593-d223f7449a82