Build/Test Tools: Improve certificate-related build scripts.
Because an exact version is pinned for `composer/ca-bundle`, the `composer update` command cannot update the dependency to the latest version. The command also does not work for a single dependency due to the fact that Composer has been configured not to generate a `composer.lock` file.

This updates the Grunt task to determine the new version using `composer outdated` before running `composer require composer/ca-bundle:NEW_VERSION --dev` to properly update the pinned version.

This also updates the build process test workflow to confirm that the certificate-related files under version control in `src/wp-includes/certificates` are up to date.

Props johnbillion.
Fixes #63939.

git-svn-id: https://develop.svn.wordpress.org/trunk@60765 602fd350-edb4-49c9-b593-d223f7449a82