Upgrade/Install: Update sodium_compat to v1.21.2.
Version 1.21.2 is mostly to prevent deprecation warnings on PHP 8.5. These should never be encountered in practice (as the only file in scope was an `SplFixedArray` polyfill for PHP 5.2, which should never be executed on PHP 8.5), but for completeness the polyfill was updated.

References:
* [https://github.com/paragonie/sodium_compat/releases/tag/v1.21.2 sodium_compat 1.21.2 release notes]
* [https://github.com/paragonie/sodium_compat/compare/v1.20.1...v1.21.2 Full list of changes in sodium_compat 1.21.2]

Follow-up to [55699], [58752], [58753].

Props paragoninitiativeenterprises, jrf, johnbillion, TobiasBg, SergeyBiryukov.
Fixes #64008.

git-svn-id: https://develop.svn.wordpress.org/trunk@60787 602fd350-edb4-49c9-b593-d223f7449a82