Cron API: Spawn cron jobs on shutdown hook.
Modifies `wp_cron()` to register the spawning of cron jobs on the `shutdown` hook rather than the `wp_loaded` hook.

The loopback request made while spawning cron jobs is intended to be non-blocking, however in some circumstances that isn't the case and the spawning will increase the time to first byte.

Props westonruter, pmbaldha, mukesh27.
Fixes #63858.


git-svn-id: https://develop.svn.wordpress.org/trunk@60925 602fd350-edb4-49c9-b593-d223f7449a82