Networks and Sites: introduce 3 new actions inside the `populate_network()` function.
* `before_populate_network`
* `after_populate_network`
* `after_upgrade_to_multisite`

These hooks are strategically located to allow the execution of custom code before & after a new multisite network is created, as well as after the first time a single-site installation is upgraded to multisite.

This change also includes a set of unit tests via a new `Tests_Multisite_PopulateNetworkHooks` class, to confirm that these hooks are executing when and as intended.

Props jeremyfelt, johnjamesjacoby, Mista-Flo, spacedmonkey, sukhendu2002, rmccue, rollybueno.

Fixes #27289.

git-svn-id: https://develop.svn.wordpress.org/trunk@60954 602fd350-edb4-49c9-b593-d223f7449a82