Upgrade/Install: Indicate use of `MyISAM` storage engine in upgrade check.
In MySQL 5.5.5, the default storage engine was changed from `MyISAM` to `InnoDB`. While still available, usage of `MyISAM` has been discouraged since and is considered legacy due to the lack of support for more modern feature. 

The percentage of WordPress sites with `MyISAM` tables in the wild is currently unknown. This change adds a field to upgrade checks that includes a list of tables using `MyISAM` to help make more informed decisions about database features in the future.

Props johnjamesjacoby, johnbillion, dd32, desrosj, mukesh27.
Fixes #63640.

git-svn-id: https://develop.svn.wordpress.org/trunk@61001 602fd350-edb4-49c9-b593-d223f7449a82