REST API: Improve performance of `X-WP-Total`/`X-WP-TotalPages` queries.
Improve the performance of fallback queries to determine the total number of objects available on various REST endpoints for out of bounds queries.

The database queries are modified to return the minimum amount of data required for determining the count and bypass priming of meta a term caches where appropriate.

Props adamsilverstein, joehoyle, johnbillion, jorbin, kadamwhite, spacedmonkey, sukhendu2002, westonruter.
Fixes #62801.


git-svn-id: https://develop.svn.wordpress.org/trunk@61002 602fd350-edb4-49c9-b593-d223f7449a82