Query: Prevent querying for all terms in `WP_Term_Query` when `include` is set to `[0]`.
This change brings the `include` parameter of `WP_Term_Query` up-to-speed with the `include` and `__in` parameters of other `_Query` classes, so that sending a value of `[0]` will not unintentionally perform an unbound query that returns all Terms.

It also introduces 3 new unit tests (for the Post, Term, and User query classes) to compare this behavior between them and ensure they remain consistent going forward.

Props audrasjb, hareesh-pillai, hellofromTonya, johnjamesjacoby, mukesh27.

Fixes #47719.

git-svn-id: https://develop.svn.wordpress.org/trunk@61048 602fd350-edb4-49c9-b593-d223f7449a82