Script Loader: Restore original return value for `WP_Script_Modules::get_dependencies()`.
Even though this method is private, there are some usages of it in the ecosystem via the Reflection API. So this reverts the script module IDs `string[]` return value in favor of the original `array<string, array>` return value. This re-emphasizes the need for more public accessor methods being tracked in #60597. 

Developed in https://github.com/WordPress/wordpress-develop/pull/10403

Follow-up to [60999].

Props pbiron, westonruter, johnbillion. 
See #63486, #60597.


git-svn-id: https://develop.svn.wordpress.org/trunk@61073 602fd350-edb4-49c9-b593-d223f7449a82