Build/Test Tools: Remove `in:body` from pull request search query.
When searching for pull requests using the GraphQL API, `in:body` is not being respected in search queries. Because it’s common for Trac tickets to be referenced in the discussions happening in comments and reviews, some pull requests are being closed out incorrectly.

This removes `in:body` from the search query introduced in [59661] and instead checks the `bodyText` for references to the fixed tickets.

The `state` field is also being removed because it is unused.

Props dd32, peterwilsoncc, johnbillion, joedolson, westonruter, dmsnell.
Fixes #64175.

git-svn-id: https://develop.svn.wordpress.org/trunk@61100 602fd350-edb4-49c9-b593-d223f7449a82