General: Clarify that contributors should fork `wordpress-develop`.
Because `WordPress/wordpress-develop` is a mirror, the only way to contribute to WordPress on GitHub is by forking the repository (even when a contributor has write access as a Core Committer). This updates the “Getting Started” section to make this more clear.

The section about how to contribute using GitHub Codespaces is also being removed since this has been broken for some time (see #57896). This section can be reintroduced once the underlying tooling is fixed.

Props pbearne, desrosj, westonruter, wildworks, peterwilsoncc.
Fixes #64100. See #57896.

git-svn-id: https://develop.svn.wordpress.org/trunk@61202 602fd350-edb4-49c9-b593-d223f7449a82