Docs: Fix phpdoc types for `WP_Speculation_Rules` members.
What was previously `mixed` is actually a speculation rule which can be more accurately be typed as `array<string, mixed>`. Additionally, the return value of `jsonSerialize` is fixed to indicate it returns an array of lists as opposed an array of associative arrays.

Developed in https://github.com/WordPress/wordpress-develop/pull/8946

Follow-up to [59837].

Props justlevine, johnbillion, westonruter.
See #64238, #63268, #62503.


git-svn-id: https://develop.svn.wordpress.org/trunk@61280 602fd350-edb4-49c9-b593-d223f7449a82